% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allyBar.R
\name{ally_bar}
\alias{ally_bar}
\title{ally_bar}
\usage{
ally_bar(data, mapping, numbers = F, angle = 40, hjust = 1)
}
\arguments{
\item{data}{data frame}

\item{mapping}{aes passed to ggplot}

\item{numbers}{display count or not, Default: F}

\item{angle}{angle passed to theme axis.text.x, Default: 40}

\item{hjust}{hjust passed to theme axis.text.x, Default: 1}
}
\value{
ggplot object (passed to ggally)
}
\description{
barchart to be used with ggally
}

