% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allyBox.R
\name{ally_box}
\alias{ally_box}
\title{ally_box}
\usage{
ally_box(data, mapping, angle = 40, hjust = 1, x.grid = F)
}
\arguments{
\item{data}{data frame}

\item{mapping}{aes passed to ggplot}

\item{angle}{angle passed to theme axis.text.x, Default: 40}

\item{hjust}{hjust passed to theme axis.text.x, Default: 1}

\item{x.grid}{panel grid on x, Default: F}
}
\value{
ggplot object (passed to ggally)
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}

