% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ally_cor.R
\name{ally_cor}
\alias{ally_cor}
\title{correlation plot for ggally}
\usage{
ally_cor(data, mapping, cor_method = "pearson",
  cor_use = "pairwise.complete.obs", cor_threshold = 0.4,
  cor_color = "red", ...)
}
\arguments{
\item{data}{data frame}

\item{mapping}{aes passed to ggplot}

\item{cor_method}{method for correlation estimation, Default: 'pearson'}

\item{cor_use}{argument to cor, Default: 'pairwise.complete.obs'}

\item{cor_threshold}{threshold for color switch, Default: 0.4}

\item{cor_color}{color used at color switch, Default: 'red'}
}
\value{
ggplot object (passed to ggally)
}
\description{
correlation plot to be used with ggally
}
\seealso{
\code{\link[vcd]{assocstats}}
}

