% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ally_count.R
\name{ally_count}
\alias{ally_count}
\title{count plot for ggally}
\usage{
ally_count(data, mapping, counts = "abs", angle = 35, hjust_x = 1,
  vjust_y = 0, ...)
}
\arguments{
\item{data}{data frame}

\item{mapping}{aes passed to ggplot}

\item{counts}{absolute number or relative, see details, Default: 'rel'}

\item{angle}{angle passed to theme axis.text Default: 35}

\item{hjust_x}{hjust passed to theme axis.text.x, Default: 1}

\item{vjust_y}{vjust passed to theme axis.text.y, Default:0}

\item{...}{PARAM_DESCRIPTION}
}
\value{
ggplot object (passed to ggally)
}
\description{
count plot to be used with ggally. To be used with categorical variables.
}
\details{
counts: one of 'abs' or 'rel'. To be expanded.
}

