% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ally_scatter_lm_cor.R
\name{ally_scatter_lm_cor}
\alias{ally_scatter_lm_cor}
\title{linear regression with correlation threshold for ggally}
\usage{
ally_scatter_lm_cor(data, mapping, cor_method = "pearson",
  cor_use = "pairwise.complete.obs", cor_threshold = 0.4,
  cor_color = "red", ...)
}
\arguments{
\item{data}{data frame}

\item{mapping}{aes passed to ggplot}

\item{cor_method}{method for correlation estimation, Default: 'pearson'}

\item{cor_use}{argument to cor, Default: 'pairwise.complete.obs'}

\item{cor_threshold}{threshold for color switch, Default: 0.4}

\item{cor_color}{colour for used at color switch, Default: 'red'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
ggplot object (passed to ggally)
}
\description{
scatter plot with linear regression and correlation threshold to be used with ggally
}
