% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aztable.R
\name{aztable}
\alias{aztable}
\title{az specific pixidust table}
\usage{
aztable(df, caption = NULL, label = NULL, print_method = "latex",
  long = F)
}
\arguments{
\item{df}{data frame}

\item{caption}{string with caption for table, Default: NULL}

\item{label}{string with label for cross-reference. "tab:" is automatically added, Default: NULL}

\item{print_method}{"latex" or "html",  Default: "latex"}

\item{long}{use longtable environment (for latex-use only), Default: F}
}
\value{
pixiedust object
}
\description{
Wrapper for az specifications of pixiedust tables
}

