% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Check for included columns}
\usage{
check_columns(.df = rawdata, columns = all_cols)
}
\arguments{
\item{.df}{dataframe to look in, Default: rawdata}

\item{columns}{vector of columns that should be included, Default: all_cols}
}
\value{
message with Yes or a list of missing columns
}
\description{
Check for included columns
}
