% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_message}
\alias{check_message}
\title{Prints a message based on a single TRUE / FALSE}
\usage{
check_message(logical)
}
\arguments{
\item{logical}{Single logical}
}
\value{
message with "Yes" if TRUE and "No" if FALSE
}
\description{
Prints a message based on a single TRUE / FALSE. Only included to be printed as
output in markdown files wihtout using print()
}
