% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_unique}
\alias{collapse_unique}
\title{Collapses unique values}
\usage{
collapse_unique(x)
}
\arguments{
\item{x}{vector}
}
\value{
character vector of length 1
}
\description{
Prints a comma separated string with unique values of a vector
}
