% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_proportionality_anova.R
\name{dose_proportionality_anova}
\alias{dose_proportionality_anova}
\title{Dose proportionality test using anova}
\usage{
dose_proportionality_anova(df, x, y, plot.par = F, max_position = 1.25,
  min_position = 0.85, signif = 3)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{x}{x variable (e.g. dose)}

\item{y}{y variable (e.g. auc)}

\item{plot.par}{print anova parameters as table, Default: F}

\item{max_position}{multiplier to max(y) for positioning text annotation, Default: 1.25}

\item{min_position}{multiplier to max(y) for positioning text annotation, Default: 0.85}

\item{signif}{PARAM_DESCRIPTION, Default: 3}
}
\value{
a list of ggplot object, summary of anova fit and the fit object
}
\description{
Anova fit to test for dose proportionality
}
\details{
Table with parameters extend between min_position*max(y) and max_position*max(y)
}
