% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_proportionality_linear.R
\name{dose_proportionality_linear}
\alias{dose_proportionality_linear}
\title{Dose proportionality test (linear model)}
\usage{
dose_proportionality_linear(df, x, y, ci.level = 0.95, plot.par = F,
  x_min = 0, max_position = 1.2, min_tab_position = 1.05,
  min_p_position = 0.9, signif = 3, p_value_threshold = 0.05)
}
\arguments{
\item{df}{data frame}

\item{x}{x variable (e.g. dose)}

\item{y}{y parameter (e.g. auc)}

\item{ci.level}{confidence level for lm fit in plot, Default: 0.95}

\item{plot.par}{add table with lm parameters, Default: F}

\item{x_min}{lower limit passed to coord_carteesian (absolute), Default: 0}

\item{max_position}{upper y limit passed to coord_carteesian (relative max(y)), Default: 1.2}

\item{min_tab_position}{lower y-value for parameter table (relative max(y)), Default: 1.05}

\item{min_p_position}{position of p-value string (relative max(y)), Default: 0.9}

\item{signif}{number of significant numbers in parameter table, Default: 3}

\item{p_value_threshold}{p-value threshold for inclusion of intercept, Default: 0.05}
}
\value{
list with plot, lm summary and lm fit
}
\description{
lm fit to test for dose proportionality
}
