% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBarchart.R
\name{gg_bar}
\alias{gg_bar}
\title{gg_bar}
\usage{
gg_bar(df, x, numbers = F)
}
\arguments{
\item{df}{data frame}

\item{x}{name of column to plot}

\item{numbers}{print count, Default: F}
}
\value{
ggplot object
}

