% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_box.R
\name{gg_box}
\alias{gg_box}
\title{boxplot}
\usage{
gg_box(df, x, y, angle = 40, hjust = 1, x.grid = F)
}
\arguments{
\item{df}{dataframe}

\item{x}{x variable}

\item{y}{y variable}

\item{angle}{angle passed to theme axis.text.x, Default: 40}

\item{hjust}{hjust passed to theme axis.text.x, Default: 1}

\item{x.grid}{panel grid on x, Default: F}
}
\value{
ggplot object
}
\description{
boxplot
}
