% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_box_continuous.R
\name{gg_box_continuous}
\alias{gg_box_continuous}
\title{box plot with x-axis spaced as continous}
\usage{
gg_box_continuous(df, x, y, fill = NULL, dodge = 0)
}
\arguments{
\item{df}{data frame}

\item{x}{x variable (continuous)}

\item{y}{y variable}

\item{fill}{fill variable passed to aes, Default: NULL}

\item{dodge}{position of text, Default: 0}
}
\value{
ggplot object
}
\description{
box plot with the x-axis spaced out as continous instead of categorical
}
\details{
x-axis is continous and then grouped, useful for dose-linearity graphics
}
