% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_count.R
\name{gg_count}
\alias{gg_count}
\title{create a counts plot}
\usage{
gg_count(df, x, y, counts = "abs", angle = 35, hjust_x = 1, vjust_y = 0,
  ...)
}
\arguments{
\item{df}{dataframe}

\item{x}{x variable}

\item{y}{y variable}

\item{counts}{absolute number or relative, see details, Default: 'rel'}

\item{angle}{angle passed to theme axis.text Default: 35}

\item{hjust_x}{hjust passed to theme axis.text.x, Default: 1}

\item{vjust_y}{vjust passed to theme axis.text.y, Default:0}

\item{...}{PARAM_DESCRIPTION}

\item{mapping}{aes passed to ggplot}
}
\value{
ggplot object
}
\description{
count plot
}
\details{
proably not working - needs debugging to det the rlang correct...

counts: one of 'abs' or 'rel'. To be expanded.
}

