% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCount.R
\name{gg_counts}
\alias{gg_counts}
\title{create a counts function}
\usage{
gg_counts(df, x, y, corThreshold = 0.4, numbers = F, ...)
}
\arguments{
\item{df}{dataframe}

\item{x}{x variable}

\item{y}{y variable}

\item{corThreshold}{correlation threshold, Default: 0.4}

\item{numbers}{whether to provide the numbers, Default: F}
}
\value{
ggplot object
}
