% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_density.R
\name{gg_density}
\alias{gg_density}
\title{density plot}
\usage{
gg_density(df, x, n_bins = 6, with_est = F, dist = NULL, var = NULL)
}
\arguments{
\item{df}{data frame}

\item{x}{x variable}

\item{n_bins}{number of bins passed to geom_freqpoly, Default: 6}

\item{with_est}{PARAM_DESCRIPTION, Default: F}

\item{dist}{pdf function for overlaying distribution, Default: NULL}

\item{var}{estimated variance of the overlaying distribution, Default: NULL}
}
\value{
ggplot object
}
\description{
density function with or without overlaying distribution
}
\details{
DETAILS
}

