% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_forest.R
\name{gg_forest}
\alias{gg_forest}
\title{forest plot for covariate-parameter impact}
\usage{
gg_forest(df, x, y, x_ci_u, x_ci_l, impact_area = c(0.8, 1.25), ...)
}
\arguments{
\item{df}{data frame}

\item{x}{x variable}

\item{y}{y variable}

\item{x_ci_u}{upper confidence interval for x variable}

\item{x_ci_l}{lower confidence interval for x variable}

\item{impact_area}{delimits shaded area, Default: c(0.8, 1.25)}
}
\value{
ggplot object
}
\description{
forest plot for covariate-parameter impact
}

