% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLmWithCor.R
\name{gg_lm_with_cor}
\alias{gg_lm_with_cor}
\title{linear regression with correlation}
\usage{
gg_lm_with_cor(df, x, y, corThreshold = 0.4, corMethod = "pearson", ...)
}
\arguments{
\item{df}{dataframe}

\item{x}{x variable}

\item{y}{y variable}

\item{corThreshold}{correlation threshold, Default: 0.4}

\item{corMethod}{correlation method, Default: 'pearson'}
}
\value{
ggplot object
}
