% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggObsVsPred.R
\name{gg_obs_vs_pred}
\alias{gg_obs_vs_pred}
\title{gg observation vs pred}
\usage{
gg_obs_vs_pred(df, y, x, grp = NULL, method = "loess", labs = list(y =
  "Observed", x = "Predicted"))
}
\arguments{
\item{df}{data frame}

\item{y}{observation variable}

\item{x}{predicted variable}

\item{grp}{group variable, Default: NULL}

\item{method}{statistical method for central tendancy, Default: 'loess'}

\item{labs}{x and y label names, passed as a list, Default: list(y = "Observed", x = "Predicted")}
}
\value{
ggplot object
}
\description{
gg observation vs pred
}

