% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggQQPlot.R
\name{gg_qq_plot}
\alias{gg_qq_plot}
\title{qq plot}
\usage{
gg_qq_plot(df, sample = CWRES, stat = "qq")
}
\arguments{
\item{df}{data frame}

\item{sample}{name of sample variable, Default: CWRES}

\item{stat}{theoretical distribution, Default: 'qq'}
}
\value{
ggplot object
}
\description{
qq plot
}

