% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_residuals.R
\name{gg_residuals}
\alias{gg_residuals}
\title{residuals vs predictions or time}
\usage{
gg_residuals(df, y = CWRES, x = PRED, absolute = F,
  smooth_method = "loess")
}
\arguments{
\item{df}{data frame}

\item{y}{y variable, Default: CWRES}

\item{x}{x variable, Default: PRED}

\item{absolute}{use absolute values of residuals?, Default: F}

\item{smooth_method}{method for smoother, Default: 'loess'}
}
\value{
ggplot object
}
\description{
plot for residuals vs predictions or time
}
