% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scatter_smooth.R
\name{gg_scatter_smooth}
\alias{gg_scatter_smooth}
\title{scatter plot with smoother}
\usage{
gg_scatter_smooth(df, x, y, smooth_method = "loess")
}
\arguments{
\item{df}{data frame}

\item{x}{x variable}

\item{y}{y variable}

\item{smooth_method}{method for smoother, Default: 'loess'}
}
\value{
ggplot object
}
\description{
scatter plot with smoother
}

