% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIndConcTime.R
\name{gg_ind_conc_time}
\alias{gg_ind_conc_time}
\title{create individual concentration time plot}
\usage{
gg_ind_conc_time(df, obs = DV, pred = PRED, ipred = IPRED, x = TAPD,
  id = ID, occ = OCC, facet_rows = 3, ...)
}
\arguments{
\item{df}{dataframe}

\item{obs}{name of observation column , Default: DV}

\item{pred}{name of prediction column, Default: PRED}

\item{ipred}{name of individual prediction column, Default: IPRED}

\item{x}{name of x variable, Default: TAPD}

\item{id}{name of ID column, Default: ID}

\item{occ}{name of OCC column, Default: OCC}

\item{facet_rows}{number of rows to factet on, Default: 3}
}

