% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_draw.R
\name{grob_draw}
\alias{grob_draw}
\title{For printing a list of grobs to pdf}
\usage{
grob_draw(x)
}
\arguments{
\item{x}{grob grids}
}
\value{
"printed" grob object
}
\description{
forces a new page for grob objects to avoid printing all on the same page
}
\seealso{
\code{\link[grid]{grid.newpage}},\code{\link[grid]{grid.draw}}
}
