% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_data_split.R
\name{ind_data_split}
\alias{ind_data_split}
\title{ind_data_split}
\usage{
ind_data_split(df, n_per_page = 12, id = "ID")
}
\arguments{
\item{df}{data frame}

\item{n_per_page}{number of ids/key variable per page, Default: 12}

\item{id}{column name of id/key variable, Default: 'ID'}
}
\value{
list with subsets of the data frame
}
\description{
splits data into list subsets for plotting
}
