% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_cor_extract.R
\name{lst_cor_extract}
\alias{lst_cor_extract}
\title{Pairwise extraction of omega correlations}
\usage{
lst_cor_extract(listobject, omegas, correlations, signif = 3, occ = F)
}
\arguments{
\item{listobject}{list object generated by read.lst}

\item{omegas}{character vector with omega parameters with correlation}

\item{correlations}{}

\item{signif}{number of significant digits for estimates etc, Default: 3}

\item{occ}{correlation from BOV parameters?, Default=F}
}
\value{
summarized data frame
}
\description{
Pairwise extraction of omega correlations read.lst object to data frame
}

