% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstParExtract.R
\name{lst_par_extract}
\alias{lst_par_extract}
\title{Extract model parameters from read.lst object to data frame}
\usage{
lst_par_extract(listobject, theta_parms, omega_parms, sigma_parms, signif = 3)
}
\arguments{
\item{listobject}{list object generated by read.lst}

\item{theta_parms}{character vector with theta parameters}

\item{omega_parms}{character vector with omega parameters}

\item{sigma_parms}{character vector with sigma parameters}

\item{signif}{number of significant digits for estimates etc, Default: 3}
}
\value{
summarized data frame
}
\description{
Extract model parameters from read.lst object to data frame
}
\seealso{
\code{\link[dplyr]{bind_rows}}
}

