% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxplore.R
\docType{package}
\name{pmxplore}
\alias{pmxplore}
\alias{pmxplore-package}
\alias{pmxplore-package}
\title{xpmxplore: Customized (gg)plots for PK/PD data}
\description{
This package contains wrappers for ggplot and ggally for exploration of 
pharmacokinetic/pharmacodynamic data
}
\author{
Helena Edlund \email{helena.edlund@astrazeneca.com}

Philip Delff \email{philip.delff@astrazeneca.com}

Devin Pastoor \email{devin.pastoor@gmail.com}
}
