% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmEq.R
\name{print_lm_eq}
\alias{print_lm_eq}
\title{regression equation and r2 to}
\usage{
print_lm_eq(lm_fit)
}
\arguments{
\item{lm_fit}{model object from lm fit}
}
\value{
parsable character string
}
\description{
function to add regression equation and R2 to plot
}
\details{
copied and modified from:
}

