% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDataStructure.R
\name{rDataStructure}
\alias{rDataStructure}
\title{create a datastructure?}
\usage{
rDataStructure(.df, dataSpec = "dataSpecificationFile.csv",
  dataSpecDir = "./Scripts", NMOutput = FALSE)
}
\arguments{
\item{.df}{data frame}

\item{dataSpec}{specification file, Default: 'dataSpecificationFile.csv'}

\item{dataSpecDir}{PARAM_DESCRIPTION, Default: './Scripts'}

\item{NMOutput}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
set up data from and object + specification file

The function reads in a .csv file with the data specifications 
(also accepts the same as a already loaded data frame)
The data spec will be delivered with the dataset if requested through QCP order form
and should include the columns: Variable, Code, DeCode, Description, R Variable Type. 
The description column may be changed if a different description is preferred
}
\details{
DETAILS
}

