% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_cols_mapping}
\alias{rename_cols_mapping}
\title{Renames data frame columns based on mapping}
\usage{
rename_cols_mapping(df, mapping_df)
}
\arguments{
\item{df}{data frame to rename}

\item{mapping_df}{data frame with 2 columns: first column is original names and second new names}
}
\value{
data frame
}
\description{
Prints a comma separated string with unique values of a vector
}
