% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshCalvin.R
\name{ssh_calvin}
\alias{ssh_calvin}
\title{ssh function from zilter to calvin}
\usage{
ssh_calvin(cmd, dir = "./", ...)
}
\arguments{
\item{cmd}{string of unix command to execute}

\item{dir}{wd where the cmd is executed, Default: './'}

\item{...}{additional to system()}
}
\value{
character string echoed from submitted unix command
}
\description{
ssh function from zilter to calvin
}
\details{

}

