% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cat.R
\name{summarize_cat}
\alias{summarize_cat}
\title{summarize a categorical variable}
\usage{
summarize_cat(df, col_name, signif = 3)
}
\arguments{
\item{df}{dataframe}

\item{col_name}{column name}

\item{signif}{number of significant digits to output}
}
\value{
summary data frame
}
\details{
DETAILS
}

