% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCont.R
\name{summarize_cont}
\alias{summarize_cont}
\title{summarize a continuous variable}
\usage{
summarize_cont(df, col_name, p = c(0.05, 0.95), signif = 3,
  missingFlag = -99)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{col_name}{PARAM_DESCRIPTION}

\item{p}{quantiles to output, Default: c(0.05, 0.95)}

\item{signif}{number of significant digits to output, Default=3}

\item{missingFlag}{flag indicating missing variables, Default=-99}
}
\value{
summary dataframe
}
\details{
DETAILS
}

