% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_run.R
\name{summarize_run}
\alias{summarize_run}
\title{summarizes nonmem run into data.frame}
\usage{
summarize_run(mod_no, ref_mod_no, listobject, record, descr)
}
\arguments{
\item{mod_no}{string with model number}

\item{ref_mod_no}{string with referebne model number}

\item{listobject}{listobject for run generated with read.lst}

\item{record}{data frame with run-record}

\item{descr}{string with model description}
}
\value{
data frame
}
\description{
summarizes nonmem run into data.frame
}

