% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCat.R
\name{summarizeCat}
\alias{summarizeCat}
\title{summarize a categorical variable}
\usage{
summarizeCat(df, colName)
}
\arguments{
\item{df}{dataframe}

\item{colName}{column name}
}
\value{
summary data frame
}
\details{
DETAILS
}
