% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCont.R
\name{summarizeCont}
\alias{summarizeCont}
\title{summarize continuous variables}
\usage{
summarizeCont(df, colName, p = c(0.05, 0.95))
}
\arguments{
\item{df}{dataframe}

\item{colName}{summarization column}

\item{p}{quantiles to summarize, Default: c(0.05, 0.95)}
}
