% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCov.R
\name{summarizeCov}
\alias{summarizeCov}
\title{summarize covariates}
\usage{
summarizeCov(df, covVector, type = "Cont", ...)
}
\arguments{
\item{df}{data frame}

\item{covVector}{vector of column names}

\item{type}{continous or categorical variable, Default: 'Cont'}

\item{...}{additional details to pass to summaryFun}
}
\value{
summary dataframe
}
\description{
summarize covariates either of type continuous or categorical
}
