% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_xpose_wrap.R
\name{vpc_xpose_wrap}
\alias{vpc_xpose_wrap}
\title{customized xpose vpc}
\usage{
vpc_xpose_wrap(vpcresultsfile = "vpc_results.csv", vpctabfile, type = "p",
  xlab = "Time after dose (h)", ylab = "Concentrations (ng/mL)")
}
\arguments{
\item{vpcresultsfile}{psn vpc output file}

\item{vpctabfile}{psn vpc output file ("vpctabXXX")}

\item{type}{with our without data, Default: 'p'}

\item{xlab}{string with x-axis label, Default: 'Time after dose (h)'}

\item{ylab}{string with y-axis label, Default: 'Concentrations (ng/mL)'}
}
\value{
lattice plot
}
\description{
wrapper for xpose vpc
}
\seealso{
\code{\link[xpose4]{xpose.VPC}}
}
