% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_package.R
\docType{package}
\name{xpmxplore}
\alias{xpmxplore}
\alias{xpmxplore-package}
\alias{xpmxplore-package}
\title{xpmxplore: Customized (gg)plots for PK/PD data}
\description{
This package contains wrappers for ggplot and ggally for exploration of 
pharmacokinetic/pharmacodynamic data
}
\author{
Helena Edlund \email{helena.edlund@astrazeneca.com}

Devin Pastoor \email{devin.pastoor@gmail.com}
}
