% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Check for included columns}
\usage{
check_columns(v1, v2)
}
\arguments{
\item{v1}{vector of column names included}

\item{v2}{vector of columns expected to be included}
}
\value{
message with "Passed" or a list of columns in diff
}
\description{
Check for included columns
}
