% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Check for included columns}
\usage{
check_columns(vector1, vector2)
}
\arguments{
\item{vector1}{vector of column names included}

\item{vector2}{vector of columns expected to be included}
}
\value{
message with Yes or a list of missing columns
}
\description{
Check for included columns
}
