% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_bar.R
\name{gg_bar}
\alias{gg_bar}
\title{gg_bar}
\usage{
gg_bar(df, x, numbers = F, angle = 40, hjust = 1)
}
\arguments{
\item{df}{data frame}

\item{x}{name of column to plot}

\item{numbers}{print count, Default: F}

\item{angle}{angle passed to theme axis.text.x, Default: 40}

\item{hjust}{hjust passed to theme axis.text.x, Default: 1}
}
\value{
ggplot object
}

