% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBoxplot.R
\name{gg_box_plot}
\alias{gg_box_plot}
\title{boxplot}
\usage{
gg_box_plot(df, x, y)
}
\arguments{
\item{df}{dataframe}

\item{x}{x variable}

\item{y}{y variable}
}
\value{
ggplot object
}
\description{
boxplot
}
