% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggConcTime.R
\name{gg_conc_time}
\alias{gg_conc_time}
\title{Concentration vs time plots}
\usage{
gg_conc_time(df, y = DV, x = TAFD, color = ID, grp = ID.OCC,
  occ = OCC, blq = BLQ, lloq = 1, lloq_col = "blue",
  lloq_type = "dashed")
}
\arguments{
\item{df}{data frame}

\item{y}{y variable, Default: DV}

\item{x}{x variable, Default: TAFD}

\item{color}{color variable passed to aes, should be factor, Default: ID}

\item{grp}{grouping variable passed to aes, Default: ID.OCC}

\item{occ}{col name for occasion column, Default: OCC}

\item{blq}{col name for blq flag column, Default: BLQ}

\item{lloq}{lower limit of quantification value, Default: 1}

\item{lloq_col}{color of lloq line, Default: "blue"}

\item{lloq_type}{linetyp of lloq line, Default: "dashed"}
}
\value{
ggplot object
}
\description{
Conc time plots with stratification
}
\details{
DETAILS
}

