% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_conc_time_ind.R
\name{gg_conc_time_ind}
\alias{gg_conc_time_ind}
\title{Individual concentration time plot}
\usage{
gg_conc_time_ind(df, obs = DV, pred = PRED, ipred = IPRED, x = TAPD,
  id = ID, occ = NULL, facet_rows = 3, facet_scales = "free", ...)
}
\arguments{
\item{df}{dataframe}

\item{obs}{observed data column, Default: DV}

\item{pred}{population predictions, Default: PRED}

\item{ipred}{individual predictions, Default: IPRED}

\item{x}{x variable, Default: TAPD}

\item{id}{id variable, Default: ID}

\item{occ}{occasion. If not NULL used as grouping variable (see details), Default: NULL}

\item{facet_rows}{rows used in facet_wrap, Default: 3}

\item{facet_scales}{scales used facet_wrap, Default: "free"}
}
\description{
Function requires model predictions. 
For plots without predictions use gg_conc_time() and add facets
}
\details{
If occ is provided, id and occ are combined using paste to create a new grouping variable for each occasion. 
Lines only connect data from each unique color+occ group. Useful for time after dose graphics.
Population predictions: dashed lines and triangles,
Individal predictions: lines and crosses, 
Observed data: open circles.
}

