% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTitlePlot.R
\name{gg_title_plot}
\alias{gg_title_plot}
\title{title page}
\usage{
gg_title_plot(title, text_size = 18)
}
\arguments{
\item{title}{test string to output in plot}

\item{text_size}{size of text, Default: 18}
}
\value{
ggplot object
}
\description{
Plot to serve as divider page for multiple plot pdfs
}

