% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indDataSplit.R
\name{ind_data_split}
\alias{ind_data_split}
\title{split dataset into sub elements}
\usage{
ind_data_split(df, nPerPage = 12, id = "ID")
}
\arguments{
\item{df}{data frame}

\item{nPerPage}{number of groups per split, Default: 12}

\item{id}{PARAM_DESCRIPTION, Default: 'ID'}
}
\value{
OUTPUT_DESCRIPTION
}
\details{
DETAILS
}
