% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCov.R
\name{summarize_cov}
\alias{summarize_cov}
\title{summarize covariates}
\usage{
summarize_cov(df, cov_vector, type = "Cont", ...)
}
\arguments{
\item{df}{data frame}

\item{cov_vector}{vector of column names}

\item{type}{continous or categorical variable, Default: 'Cont'}

\item{...}{additional details to pass to summary_fun}
}
\value{
summary dataframe
}
\description{
summarize covariates either of type continuous or categorical
}

