% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCont.R
\name{summarizeCont}
\alias{summarizeCont}
\title{FUNCTION_TITLE}
\usage{
summarizeCont(df, colName, p = c(0.05, 0.95))
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{colName}{PARAM_DESCRIPTION}

\item{p}{PARAM_DESCRIPTION, Default: c(0.05, 0.95)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}

